requirements.txt
appdirs==1.4.2
beautifulsoup4==4.5.1
bleach==1.4.3
chardet==2.3.0
Django==1.8.18
django-cleanup==0.4.2
django-contrib-comments==1.7.3
django-debug-toolbar==1.6
django-private-storage==1.0.2
djangorestframework==3.5.3
ecdsa==0.13
Fabric==1.12.1
filebrowser-safe==0.4.6
funcsigs==1.0.2
future==0.15.2
grappelli-safe==0.4.5
html5lib==0.9999999
Mezzanine==4.2.2
mock==2.0.0
oauthlib==2.0.0
packaging==16.8
paramiko==1.17.2
pbr==1.10.0
Pillow==3.4.2
psycopg2==2.6.2
pycrypto==2.6.1
pyparsing==2.1.10
pytz==2016.10
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
sqlparse==0.2.2
stripe==1.46.0
tzlocal==1.3
